few handlers added, signup/login workflow
