+++ templates/signin_signup.html
@@ -6,7 +6,7 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
         <meta name="description" content="">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>Sign in/Sign up</title>
+        <title>bartergo</title>
         <meta name="description" content="Item">
         <meta name="author" content="btshopng">
 
@@ -24,27 +24,23 @@
       <!-- main nav -->
         <nav class="bg-white shadow-4 cf fixed w-100 h3 flex flex-row items-center">
           <div class="w-100 db ph5 cf">
-            <div class="fl v-mid dib">
-              <a class="link black  dib ph3 dim" href="#">
-                  Btshopng
-              </a>
-            </div>
+            <div class="fl v-mid dib"><a class="link black dib ph3 dim" href="#">Btshopng</a></div>
             <div class="fr v-mid dib">
-              <form class="dib v-mid">
-                <input type="text" class="pa2" placeholder="search eg. cars, phone, etc"/>
-                <button class="link bg-black white  dib v-mid pa2 bw0">
+              <form class="dib v-mid" method="GET" action="/search">
+                <input type="text" name="searchItem" class="pa2" placeholder="search eg. cars, phone, etc"/>
+                <button type="submit" class="link bg-black white  dib v-mid pa2 bw0">
                   <img src="/public/img/search.svg" style="width:18px" class="dib"/>
                 </button>
               </form>
-              <a class="link black  dib ph3 dim" href="/">Home</a>
-              <a class="link black  dib v-mid ph3 dim" href="#">
+              <a class="link black dib ph3 dim" href="/">Home</a>
+              <a class="link black dib v-mid ph3 dim" href="/signup">
                 <img src="/public/img/user.svg" style="width:32px" class="dib ph2"/>
-                <span class="dib">login/<br/>signup</span></a>
+                <span class="dib">login/<br/>signup</span>
+              </a>
             </div>
           </div>
         </nav>
         <!-- End main nav -->
-
 
         <!-- Beginning if explore section -->
         <section class="cf vh-100 flex flex-row items-center justify-center">
